yocto: force shallow git clones
Summary:
We keep mirrors of external source artifacts so that we do not need
to access remote servers during builds.  Due to the default way Yocto
manages these mirrors, any change to the upstream repository causes
the artifact to be out of date.  Packages in the LF-OpenBMC org
regularly change due to the way Gerrit synchronizes certain state
with Github and this churn could cause our internal source mirror to
grow rather large.

Force the flags that ensure we only do shallow clones in order to
reduce both the size and churn of the internal mirror.

We have also seen periodic build failures with Kirkstone based
machines where Yocto fails to find the glibc source in the mirror.
It is hoped that forcing these shallow-clone settings, one of which
is partially enabled already for glibc, will eliminate the build
fetch failure.

Test Plan:
Updated manifold with the shallow-clone artifacts and will
check CIT results for all platforms backed to Kirkstone or above.

Reviewed By: doranand

Differential Revision: D37198271

fbshipit-source-id: 400e6ab118a0d679eda19b242960fb856520896e