rackmon2: pyrmd: Use native sync methods instead of wrapping async methods
Summary:
While great for async programming, it actually adds a lot of overhead.
For example, the average latency of a single naive sync transaction
(A dumb wrapper run on every single async call to convert it into sync)
is 85ms. Even if we convert it into an optimal async code (async method
which does N calls and we await that) is 75ms.

In contrast, if we use native sync operations end-to-end, the average
latency drops to 34ms.

This change, converts us to making sync calls use the native sockets
API instead of wrapping around the async calls. This significantly
improves the firmware upgrade time (from ~40min to 14min).

Test Plan: Firmware upgrade

Reviewed By: garnermic

Differential Revision: D37155748

fbshipit-source-id: b045a50327b084275cce7ca51452464a5e2dc287