rackmon2: Install ORv3 register map for wedge400+
Summary:
This adds conditional installation of ORv3 register
map configuration only for wedge400+ platforms. It
installs them by default but conditionally disables
installation on wedge100.

It is technically harmless on wedge100, but is an
unnecessary drain on its resources since we do
not plan to support ORv3 racks using a wedge100.

Test Plan:
Build wedge100 and wedge400 (enabling rackmon2)
and ensure the images for wedge100 does not contain
the new files and only the wedge400 contains the
newly installed files in /etc/rackmon.d

Reviewed By: garnermic

Differential Revision: D37226638

fbshipit-source-id: 9d8164a46eb36dfa295febfeb4036f53ae3312f7