common:ipmid-v2: JSON file parsing support
Summary:
config.json defines all platform-specific configuration data which are used by ipmid-v2.

The ipmid-v2 is a redesign of ipmid with the goal of making the new NPI introduction easier and reducing the code maintenance cost.
The ipmid-v2 is changed to be platform-specific configuration driving without PALLIB support.

Test Plan:
read and parse this configuration with cJSON API.
check each configuration item value with the internal presentation in the ipmid-v2.

Reviewed By: tao-ren

Differential Revision: D37705499

fbshipit-source-id: dd409084ac64c2d469fe6cd6ce2a5596169ed65d