common:ipmid-v2: porting the ipmid code to ipmid-v2
Summary:
This diff is porting of the original ipmid code.

The ipmid-v2 is a redesign of ipmid with the goal of making the new NPI introduction easier and reducing the code maintenance cost.
The ipmid-v2 is changed to be platform-specific configuration driving without PALLIB support.

Test Plan:
The ipmid-v2 is tested by running ipmitool on the X86 side.  ipmitool will communicate with ipmid-v2 in BMC side.
tested various ipmitool subcommands:
1. ipmitool  mc info
2. ipmitool fru
3. ipmitool  mc guid
4. ipmitool raw

tracing the ipmi messages on the stack: ipmid/kcsd on the BMC side.

Reviewed By: tao-ren

Differential Revision: D37705504

fbshipit-source-id: 3740a58a236d85d17fa411d883d2d0a16d496c64