common:ipmid-v2: common supporting functions for ipmid
Summary:
The common supporting functions for ipmid-v2.

The ipmid-v2 is a redesign of ipmid with the goal of making the new NPI introduction easier and reducing the code maintenance cost.
The ipmid-v2 is changed to be platform-specific configuration driving without PALLIB support.

Test Plan:
The ipmid-v2 is tested by running ipmitool on the X86 side.  ipmitool will communicate with ipmid-v2 in BMC side.
tested various ipmitool subcommands:
ipmitool  mc info
ipmitool fru
ipmitool  mc guid
ipmitool raw
tracing the ipmi messages on the stack: ipmid/kcsd on the BMC side.

Reviewed By: tao-ren

Differential Revision: D37705502

fbshipit-source-id: a182d7917b581828388731088667c9b29105087e