common: Support debug card show DIMM loop pattern (#2273)
Summary:
Summay:
Add a frame for debug card to show the decoded DIMM loop pattern.
We keep monitoring post code to detect if DIMM loop occurs, and when it happens we will set the key. Because the DIMM loop is a dead loop, the key only be cleaned when user does AC-cycle or the slot was hot-swapped.

Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/2273

Test Plan:
Build and test pass on fby35 system.
1. Successfully show the DIMM loop patter on LCD debug card
(1) Two slots occur DIMM loop
(2) Two DIMM occur DIMM loop in a slot

Reviewed By: peterdelevoryas

Differential Revision: D37820008

Pulled By: garnermic

fbshipit-source-id: 8fe5bbdfb8b2f7a4ce5d35357ee9f65c703eef10