tpm-factory-upd: Allow deprecated OpenSSL API's
Summary:
In OpenSSL 3.0, directly using the HMAC functions has been deprecated,
and code should be migrated to EVP (selecting HMAC as a parameter).
Similarly `RSA_*` functions are supposed to be migrated to `EVP_PKEY_*`
functions.

We should also ask Infineon to update this code to support OpenSSL 3.0
and pull that in instead. It should compile with `#define
OPENSSL_API_COMPAT 0x30000000` and `#define OPENSSL_NO_DEPRECATED`.

Test Plan: N/A

Reviewed By: williamspatrick

Differential Revision: D38492847

fbshipit-source-id: 73a33fc17a48d081e03c6e9a3ed8a3634b017b4a