grandcanyon: Migrate MD5 code to OpenSSL 3.0
Summary:
The MD5 (https://github.com/facebookexternal/openbmc.wiwynn/commit/abea4bf7bf4164b69571fb220231a134115786f5) functions were deprecated in OpenSSL 3 and need to be updated to use the EVP digest functions. It's a one-to-one change for the most part.

I also noticed that the code was using "strncmp" (clearly not correct right? If there's a null-byte in the digest it will short-circuit) and removed some unnecessary casting by switching to `unsigned char`.

Test Plan: N/A

Reviewed By: williamspatrick

Differential Revision: D38492845

fbshipit-source-id: 5bf8607182fa42b0f83e26ddfd83c287b4ac0ff2