qemu: Update slirp revision
Summary:
I was pointing to the version of slirp that QEMU points to right now,
but actually we'd like to be pointing at a newer revision that includes
my NC-SI changes.

Unfortunately, while I have the changes merged in libslirp, I still
haven't managed to get those changes merged into QEMU itself because the
version handling is a mess and requires relatively complicated changes
to support backwards compatibility, and barely anybody has done this
before (so no precedent).

Test Plan: N/A

Reviewed By: williamspatrick

Differential Revision: D38598458

fbshipit-source-id: e9858742ee34e8d5067c4e3dde8cbecf5e02d3c7