elbert: Remove elbertvboot machine
Summary:
Right now, elbertvboot shares `local.conf.sample` with elbert, and
requires users to manually edit the configuration file to switch
`MACHINE` to `elbertvboot`.

This conflicts with the new LF OpenBMC layer we're adding which can't
be named `fby35`, because that would conflict with the fb-internal
fby35, but needs to build with `MACHINE=fby35`. This causes the `lf-obmc download-mirror` tool to not work.

This diff removes the elbertvboot machine. If we start work on this again,
we can use this diff as a starting point.

Test Plan: Verifying that elbert builds still work.

Reviewed By: zhdaniel12

Differential Revision: D38670334

fbshipit-source-id: 5c933de4a9000ebf700dc5291cf31c5834bc4743