greatlakes: Create layer to build LF OpenBMC image
Summary:
To support our upstream efforts, we can stage temporary fixes in this
recipe, and run continuous builds and tests against this image. It also
integrates with our custom version of QEMU more closely.

`ipmitool` was conflicting with the openembedded recipe, so I changed
the recipe to only append things for machines using the `openbmc-fb`
distro override, which should be fb-only machines.

We settled on using `greatlakes` for the layer and image name. The LF codebase
has been updated so that all `fby35` code is now `greatlakes`.

Test Plan:
```
source openbmc-init-build-env greatlakes build
bitbake greatlakes-image
```

Reviewed By: williamspatrick

Differential Revision: D38609721

fbshipit-source-id: 2bdd101dffe2d3fe175e252b05aa52a9d5ae7431