qemu: Add fby35-sb-cpld
Summary:
Adds a device to emulate some of the behvaior of the CPLD on the bus with the BMC and BIC (the server board CPLD, not the baseboard CPLD).

I wrote this a while ago but didn't get it upstreamed, and forgot to put it in the github repository, so I'm finally pulling it in now as part of some BMC <-> BIC testing.

The qemu-devel thread, for reference: [thread](https://lore.kernel.org/qemu-devel/20220630045133.32251-13-me@pjd.dev/)

Test Plan: N/A

Reviewed By: GoldenBug

Differential Revision: D38898945

fbshipit-source-id: 29e586c39ec5c7fb2ebcc906f1c596bc62104f20