qemu: Connect BMC with slot1 BIC in fby35
Summary:
This includes several patches to allow the BMC and BIC to send IPMB
messages to each other in the fby35 multi-SoC machine where the BMC and
BIC are both running in parallel.

Even though the BMC and BIC are connected though, this still doesn't
quite work because each SoC sends messages in master mode synchronously.

That's ok when the BMC sends messages to the BIC, but not when the BIC
sends to the BMC, because BMC is using the byte-buffered I2C mode, so
the BIC needs to wait for the BMC to ack each byte that is sent.

Test Plan: N/A

Reviewed By: GoldenBug

Differential Revision: D39006069

fbshipit-source-id: c00c175c8a6a28fd088229068fb3ebde2a705880