rackmon2: Fix rare unit test failure.
Summary:
Yeah, dont write unit tests which depend on timing
or system load. But thats the only way to test this
one unit after which all things depending on it
can be tested using mocks.

So, make the expectation between(2,3) instead of exactly
2. This allows us wiggle room when the test took just
a bit longer to stop and we had an additional call.

Test Plan: All unit tests pass

Reviewed By: peterdelevoryas

Differential Revision: D39297407

fbshipit-source-id: 7494f786db539fdd42fc35bc0f5a5c4229fda135