grandteton: Add capability for fw-util to upgrade HGX components
Summary:
This just adds a Auto endpoint which lets HMC decide based on
the image which components are being upgraded.

TODO to add other HGX components so we can get versions.

Test Plan:
Build on grandteton and load on qemu.
Ensure calling with a dummy package results in a HTTP error
(Is calling into libhgx)

Reviewed By: williamspatrick

Differential Revision: D39317941

fbshipit-source-id: df40b0ff5bdd2c89a1d8185139c34b1ae0431413