libmpsse: split libftdi-1.4 implementation
Summary:
libftdi-1.5 (which is in upstream lf-openbmc master branch
now) adds implementations for a few functions which were not present
in libftdi-1.4 and were copied into libmpsse.  The implementation in
libmpsse was almost identical to the one added in libftdi-1.5 (and
that was likely the original source for their code).

Split the duplicated functions into their own C part which can be
enabled or disabled using a meson option.  Enable it everywhere
libmppse is sold for the time being, but we will disable it in
'master' when the Yocto bump comes through.

Test Plan: Built yv35 and halfdome.  yv3 will be covered by CIT.

Reviewed By: garnermic

Differential Revision: D39986264

fbshipit-source-id: d1f643d1a3d8895d2879a26a44f1798e47586bc9