pypartition: delete it
Summary:
pypartition has been replaced by flashy for over a year now
and is not pulled into any machine recipes.  Remove it to prune the
dead code from the repo.

Test Plan:
Checked for references to `pypartition` in our code and
found none outside of comments.  There are a number of references
in fbcode but they point to an fbpkg that hasn't been updated in
almost a year, so it isn't being automatically built from this code.

Reviewed By: garnermic, doranand

Differential Revision: D38286116

fbshipit-source-id: 31362a7e3d3589d81d3604be6e5398d6916a287d