obmc-sensors: Remove syslog on duplicated label
Summary:
We generally add a label for sensors we care about. For
sensors we do not care for, we leave it with its default
label. This will for sure cause a "duplicate label" warning
and can be excessive especially for utilities which can
start and stop.

Hence remove it.

Test Plan: Build openbmc

Reviewed By: peterdelevoryas

Differential Revision: D40558149

fbshipit-source-id: 29aa6fe3d58bc1c97df6052a0fb1983f40c09b8d