qemux86: fix layer.conf file
Summary:
Some of the layer.conf variables were too aggressively changed to
the new Yocto override syntax when they should not have been, since
layer variables are not overrides.  This worked on Dunfell because
of some blind regex implementation that was backported upstream for
attempted backwards compatibility but doesn't match the real
implementation in Kirkstone.

Test Plan: Tested out qemux86 on kirkstone.

Reviewed By: peterdelevoryas

Differential Revision: D41182726

fbshipit-source-id: de3f237a9524a5a910d7bef78cb158e18b7de98d