rackmon2: Prevent server to hang because of a misbehaving client
Summary:
This adds socket timeouts to ensure that the service
does not hang with a misbehaving client.

Test Plan:
All unit tests pass.
Flash on wedge400 without the pyrmd fix (next) and ensure we recover
with error
```
rackmond[16841]: Failed to receive message
```
And after that `rackmoncli list` works normally and the svc recovers.

Reviewed By: peterdelevoryas

Differential Revision: D41633469

fbshipit-source-id: f7d3d3557f45965d062d8bf3d97ff598349754de