minipack: bump to linux kernel version 6.0
Summary: switch to the kernel 6.0.2 for BMC image.

Test Plan:
uname -r checked the kernel version
check the dmesg and lsmod to make sure no kernel boot errors and expected kernel modules are loaded.
check the sysfs for FMC and flash SPI devices and drivers are bounded.  include the details in P563405763
pass the CIT tests2 testing in BMC.

Reviewed By: tao-ren

Differential Revision: D41625573

fbshipit-source-id: f83560849a36eb85a412c0fbcf17ddf4da7c8bd9