fby3: add missing ocpdbg fru
Summary:
It was reported that running `fw-util ocpdbg --version`
resulted in a spurious "pal_get_fru_id() wrong fru ocpdbg" message
in `/var/log/messages`.  Add the ocpdbg card to the list of accepted
FRUs and assign it a proper FRU ID.

Test Plan: Ran in qemu and observed the message is gone.

Reviewed By: garnermic

Differential Revision: D42037478

fbshipit-source-id: e42abf8ca79777f843dc76a076caa46f223b0972