fw-util: do not continue on flock failure
Summary:
We call a plat-util function `single_instance_lock_blocked`
but ignore the result from it.  In some cases this function will
return a -1, indicating that we did not successfully lock.  Do not
blindly continue the operation.

In the scenario that one firmware update is on-going and someone
attempts another, we usually block.  If someone were to cause a
SIGHUP during the block, we would end up continuing along and
potentially starting two flash operations at the same time.

Test Plan: Trivial `return -1` inserted in obvious error path.

Reviewed By: peterdelevoryas

Differential Revision: D42055288

fbshipit-source-id: e2a6608f73442188582d684b4168a5f1e53ab422