openbmc: greatlakes: Rename LF code to meta-greatlakes-lf
Summary:
Rename the current greatlakes platform to greatlakes-lf	so the
Meta code can be pulled into the greatlakes layer.  To do this,
the name of the current LF code will be changed to be greatlakes-lf.
This necessitates changing the meta-facebook/meta-greatlakes layer to
be meta-facebook/meta-greatlakes-lf for	LF code.

Test Plan:
Tested and verified build with QEMU boots to boot prompt, login	is
successful and basic commands can be executed.

Successfully tested GreatLakes, Grand Teton  and YV3.5 platforms

Reviewed By: williamspatrick

Differential Revision: D42114418

fbshipit-source-id: 83def9bc6c3d0801c356e257fc8570fba0dc038c