rackmon2: Fix infinite loop if server closes socket before data
Summary:
If the server just dies before writing a response, the socket
is considered half-closed and socket.recv() will return b''.
This will cause an infinite loop where rest can sit there and
spin. Fix this by checking and retrying.

Test Plan:
Build for wedge400, copy and install new binaries.
Test for 2 days to see if the issue is reproducable.

Reviewed By: williamspatrick

Differential Revision: D42868606

fbshipit-source-id: bfc84f4b77680af0737b5ff027e1dd5f6b4013b7