flashy: let systemd handle watchdog on LF
Summary:
LF-based images allow systemd to handle the watchdog, so
there is no reason to pet it directly.  Attempting to do so causes
a warning message since the dev entry is already held and cannot be
opened.  Skip the watchdog petting operations on LF-based machines.

Test Plan: Observed warnings are not present on Bletchley.

Reviewed By: doranand

Differential Revision: D43684213

fbshipit-source-id: a01e329f62c8da3b46f35eaba89c7f388285af60