fixing sup test sensor range failure
Summary:
yamp sup test sensor is failing because the sensor reading is
around 14 to 15C which is below the range that we were testing for. I
checked with Thermal team and they confirmed that getting a value of 14
to 15C is still valid. So this code is modified to expand the range.

Test Plan:
Failure before fix on the conveyor node

Test log after this diff: P643446971

Reviewed By: tao-ren

Differential Revision: D43857799

fbshipit-source-id: d050e9c967fdd6e957325bfa655d7aab3053636d