rackmon2: Use register level locking instead of a list level
Summary:
Goal is to allow for clients to retrieve monitoring data without having to wait excessively long for registers to get updated which can take seconds to complete.

This is accomplished by making the lock more fine-grained by moving it from protecting the register list to the register history itself.

Next in order to provide atomic updates of the registers without locking the register for an excessive period of time (10-100s of ms), we use a shadow buffer to read the registers and swap active/inactive buffers after read.

While we had minimal impact to the 90-p and below numbers (within error margin) our 99-p and max numbers are improved by a factor of 50%.

Test Plan:
1. all unit tests pass.
2. Flash on Test ORv2 system and wait ~30min for all the register histories to be read (This is done for both base and test)
2. Ensure that monitoring is fine on the new codebase. Use P646142592 script (might check it into the codebase after cleanup) to ensure that all our data is never more stale than 5min (Ensure that the monitoring logic is working fine).
3. Performance test of the rest interface to ensure that we fixed the max/99% numbers
```
| type | min  | median | 90-p | 95-p | 99-p | max  |
----------------------------------------------------
| base | 0.47 | 1.8    | 2.8  |  3.8 | 17.4 | 21.7 |
| test | 0.43 | 2.0    | 3.0  |  3.5 | 6.0  | 9.7  |
```

Reviewed By: williamspatrick

Differential Revision: D41477492

fbshipit-source-id: d695c94cafc2f337e0be76f83babb2a38c1aa8a5