linux-6.1: move AST2500 patch to override
Summary:
When porting patches from 6.0 to 6.1, I included a patch
which is only applicable to the AST2500 chip.  When loading that
patch on an AST2600 machine (such as Bletchley) it will hang the
kernel due to an invalid memory access.  Switch the patch declaration
to use the override syntax instead so that it is only picked up on
AST2500-class chips.

Test Plan: Build successfully.

Reviewed By: garnermic

Differential Revision: D44094773

fbshipit-source-id: fce0fddadd45675464c8a326bb82b5549e622f48