flashy: validate: allow parts to not be present if IGNORE type
Summary:
The LF images are truncated to only the data written,
like our images, but the images can be up to 128 MB.  It is expected
for the rwfs partition to be in the image manifest but not be in
the image file.

Modify the partition parsing to not fail if a partition has a type
of 'IGNORE' and is not present in the image.  In that case, return
an empty byte slice.

This is similar behavior to the existing case where the partition can
go off the end of the image and we return a truncated slice.

Test Plan: Ran flashy in QEMU on Bletchley.

Reviewed By: doranand

Differential Revision: D44593878

fbshipit-source-id: edd5148b2f9e6a31ac281d24b02ae4b24f78dde2