ntpdate: remove all uses
Summary:
The latest version of Yocto is eliminating the ntpdate package.  There
is possibility of calling `ntpd -qgx <server>` directly for a similar
purpose, but we cannot have the ntpd running from init in that case.

The primary purpose of the ntpdate calls was to be able to prime the
NTP configuration with the values of `kv get ntp_server persistent`.

Rework the config file and associated scripts so that we can instead
update the `/etc/ntp.conf` with the server from the KV entry, if
necessary.

Test Plan:
Ran the code on Halfdome in QEMU.  Copy/paste to other platforms.

```
root@bmc-oob:~# grep "REPLACE" /etc/ntp.conf
server time.google.com iburst # REPLACE WITH KV
root@bmc-oob:~# kv get ntp_server persistent
time.google.com
```

Reviewed By: amithash, zhdaniel12

Differential Revision: D44681758

fbshipit-source-id: 0be669d7cdca7f41d8ab6df89816a4571ccef995