Upgrade vbs struct to version 1
Summary:
vbs strcture add
1. vbs_ver and crc2 to add vbs structure versioning
2. giu related three fields giu_mode, op_cert, op_cert_size

With versioning support in vbs strcture, client parsering the strcture
can be backwards compatible.

Add crc2 to cover new fields is to reduce the chance taking the
not used vbs_ver on legacy systems as valid value. Essentially is to
identify it is a legacy system or new system support versioning.

Test Plan:
Test on YV3.5 and YV3 platform with
```
   SPL(GI)    U-Boot(main)               vbs cmd success
       0              0                    no VBS version output
       0              1                      VBS version: Legacy
       1              0                    no VBS version output
       1              1                      VBS version: 1 and GIU information
```

Reviewed By: amithash

Differential Revision: D45069730

fbshipit-source-id: c47db987d109fc4c9cd6e471d0d58d26debc5f68