fuji: PIM UCD I2C hang workaround (#1490)
Summary:
We are facing UCD hanging the bus on PIMs, the rationale is not clear. Hence, we decide to remove UCD and MPS VR since they share the same i2c mux channel. Removing UCD and keeping MPS VR means UCD still sees the i2c transactions, though these transactions are not meant for UCD. There is still the possibility that UCD gets messed up.  For the systems which are suffering i2c hanging in the data center, this workaround can restore the access to other sensors on the PIMs.

The side effect is loss of monitoring of PIM onboard POL voltages, including FPGA, retimer, and QDD voltages. As for the QDD voltages, there's a substitute. QDD module voltage can be read through the QDD i2c interface (to DOM FPGA) directly.

**Summary:**
- remove PIM's UCD, and MPs from cit i2c test
- remove PIM's UCD, and MPs from cit sensor test
- remove PIM's UCD, and MPs from rest-api
- remove PIM's UCD, and MPs from sensors
- remove PIM's UCD, and MPs from front-paneld
- remove PIM's UCD, and MPs from pal

X-link: https://github.com/facebookexternal/openbmc.celestica/pull/1490

Test Plan:
- check whether `sensors` is able to work fine
- check whether `pal` and `front-paneld` are able to work fine by checking the command `sensor-utils all`  and `sensor-utils --force all`
- check whether `cit i2c test`,  `cit sensor test` and `rest-api` work fine by `cit test`

_note:
The test plan just shows the change does not impact the related system.
Because the root cause is not clear and very hard to be reproduced.
We expect If UCD and MPs are removed, this issue would not occur again._

[sensors.txt](https://github.com/facebookexternal/openbmc.celestica/files/11176923/sensors.txt)
[sensor-util.txt](https://github.com/facebookexternal/openbmc.celestica/files/11097804/sensor-util.txt)
[CIT_test.txt](https://github.com/facebookexternal/openbmc.celestica/files/11097806/CIT_test.txt)

Reviewed By: joancaneus

Differential Revision: D44769019

Pulled By: binhuang00

fbshipit-source-id: 7e7388a2dffc2584d1887d7482759d5b6b2575dc