Add back "fault_tolerant" setting to fuji's fsc-config.json
Summary:
[FUJI] Add back "fault_tolerant" setting to fuji's fsc-config.json
Some vendor made a mistake to remove fault_tolerant setting from Fuji's fsc-config.json.
This happend when that vendor changed TH4 sensor at 0x4c to 0x49. But this "fault_tolerant"
setting is about another device at 0x47. So this should not be removed. This caused a
hiccup in Meta's datacenter.

Test Plan: Tested live in one of the FUJIs in AI cluster.

Reviewed By: binhuang00

Differential Revision: D45616464

fbshipit-source-id: bd51026c57e6ca269d20193c58fb7b20ea5dca95