halfdome: Olympic 2.0: Add inlet sensor correction table (#2618)
Summary:
Add inlet sensor correction table for Olympic2.0.

Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/2618

Test Plan:
1. Build and test pass on OLP2.0 system. pass

2. Test by our Thermal team. pass.

# Description

Please include a summary of the change and which issue is fixed.

# Motivation

Please include an explanation of why you these changes are necessary

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce.

Reviewed By: williamspatrick, wangx6f

Differential Revision: D45805868

fbshipit-source-id: 97bb13b8324a7b44c251497f441ecbdbbfd5874f