common: create package group for tests2 requirements
Summary:
The LF builds by default do not include python3, but we
need to have it and a few dependencies in order to run the tests2
framework.  Create a common package group and add it to the image
include file for all LF-based machines.

Test Plan:
Built and booted in QEMU:
```
root@bmc:~# python3
Python 3.11.2 (main, Feb  7 2023, 13:52:42) [GCC 12.2.0] on linux
Type "help", "copyright", "credits" or "license" for more information.
>>> import json
>>> import unittest
>>> import syslog
>>> quit()
```

Reviewed By: jagpalgill

Differential Revision: D46167425

fbshipit-source-id: 50d94a51968021f2eb8943185009994169cdb976