fby35: gl: Fixed "Couldn't identify the system type" when reading sensor (#2635)
Summary:
Description:
Add "if" statement to determine the slot type. If it is GreatLakes, set the machine type to "Config B".

Motivation:
There's no statement for checking system type "GreatLakes", so BMC add error message to SEL.

Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/2635

Test Plan:
Before:
  root@bmc-oob:~# power-util sled-cycle
  root@bmc-oob:~# cat /var/log/messages | grep "Couldn't identiy the system type"
  2018 Mar 19 02:16:41 bmc-oob. user.warning greatlakes-v2023.20.1: sensord: read_snr_from_all_slots() Couldn't identiy the system type: Type_GL

  After:
  root@bmc-oob:~# power-util sled-cycle
  root@bmc-oob:~# cat /var/log/messages | grep "Couldn't identify the system type"
  root@bmc-oob:~#

# Description

Please include a summary of the change and which issue is fixed.

# Motivation

Please include an explanation of why you these changes are necessary

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce.

Reviewed By: williamspatrick, anitazha

Differential Revision: D46100100

fbshipit-source-id: b9365669b46b8df79dfaa7b57e9d43d6399af3be