Meta common liteon psu fw upgrade 32bytes 16bytes dev
Summary:
**Description**
Currently, Liteon PSU is supporting 1 + 1 firmware upgrades only, the PSU requires ac of the other PSU to be able to recover the unit. However, this method is not flexible to upgrade the firmware in data centers such as wedge400/wedge400c. Hence, we need to support 1 + 0 firmware to enhance firmware upgrade quality to be more flexible.

Before, the psu-util supported 16 bytes + header for firmware data transmitting. However, it takes slowly for Liteon All uc firmware upgrade ( ~10 mins ). Hence, we decided to use 32 bytes + header for the transmitting to save upgrade times, we can decrease times to 6.25 mins. Unfortunately, the SMBus API we are using, can not support 32 bytes + header for transmitting. Hence, change to the i2c protocol for firmware data transmitting ( cmd 0xF3 and 0xF2 ) instead of the SMBUS protocol.

**Summary:**

- use i2c protocol instead of smbus protocol to support 32 bytes for Liteon's fw data transmitting ( cmd 0xF3 and 0xF2 )
- support firmware version showing for all uC
- support uC all for set bootloder mode error checking, FW identifier error checking, and CRC6 application checksum checking in DEBUG mode
- support backward compatible

**Bonus fix:**
To fix FW Identifier Error and CRC16 Application Checksum are not tested for the debug mode, move the test to be before resetting the unit because the test is not available after resetting the PSU.

**Test plan for Liteon:**

- Liteon DC PSU that support 1 + 0 fw uprgrade
- Liteon DC PSU that support 1 + 1 fw upgrade ( backward compatible )
- Liteon AC PSU that support 1 + 1 fw upgrade ( backward compatible )

**Test plan for Delta:**
To confirm this change will not impact to Delta fw upgrade

- Delta DC PSU that support 1 + 0 fw uprgrade
- Delta AC PSU that support 1 + 1 fw upgrade

Logs:

[Liteon DC PSU that support 1 + 0 fw uprgrade.txt](https://github.com/facebookexternal/openbmc.celestica/files/11542021/Liteon.DC.PSU.that.support.1.%2B.0.fw.uprgrade.txt)
[Liteon DC PSU that support 1 + 1 fw upgrade.txt](https://github.com/facebookexternal/openbmc.celestica/files/11542041/Liteon.DC.PSU.that.support.1.%2B.1.fw.upgrade.txt)
[Liteon AC PSU that support 1 + 1 fw upgrade.txt](https://github.com/facebookexternal/openbmc.celestica/files/11542027/Liteon.AC.PSU.that.support.1.%2B.1.fw.upgrade.txt)
[Delta DC PSU that support 1 + 0 fw uprgrade.txt](https://github.com/facebookexternal/openbmc.celestica/files/11542032/Delta.DC.PSU.that.support.1.%2B.0.fw.uprgrade.txt)
[Delta AC PSU that support 1 + 1 fw upgrade.txt](https://github.com/facebookexternal/openbmc.celestica/files/11542037/Delta.AC.PSU.that.support.1.%2B.1.fw.upgrade.txt)

X-link: https://github.com/facebookexternal/openbmc.celestica/pull/1541

Reviewed By: binhuang00

Differential Revision: D46102897

Pulled By: mikechoifb

fbshipit-source-id: 7a641b877ea39d06dcf9f9e1e22c8c0971797fc4