spec: remove 'owners'
Summary:
Remove the explicit owner from the spec and redirect to the
compute team as a whole.  These assignments are only suppose to last
until the document is complete and they also tend to get out of date.

People in the know probably know who the POC is for a system.  People
not in the know should contact the team on-call anyhow.  This gets
them into that practice.

Test Plan: Generate spec and inspect.

Reviewed By: amithash

Differential Revision: D47036403

fbshipit-source-id: e3d401ca58fa4c56623cd9089dd6b166ab867432