spec: invert feature YAML vs parsing feature markdown
Summary:
Previously the script would use the feature YAML to generate
a FEATURE.md (if it didn't exist) and then reinterpret the FEATURE.md
to create the final markdown/docx.  This ended up being a lossy
translation and lead to awkward naming of features, due to file naming
requirements.  Invert this relationship such that the FEATURE.md is
interpreted into a [simpler] feature YAML if it is the only information
available, otherwise use the original feature YAML as a source of truth.

Test Plan:
Generate both greatlakes and bletchley spec docs.  Bletchley
is 100% identical, while Greatlakes is strictly improved.

Reviewed By: tao-ren

Differential Revision: D47317863

fbshipit-source-id: fa7286aa645d94696d0a5b0e760d88696f1d9d95