rackmon2: Fix use-after-scope-end during test runs
Summary:
This was caught by Address Sanitizer. Essentially we
have the `mon` object which uses the `commandTimeout`
bool. When the function scope is being terminated,
before this change commandTimeout is destroyed and then
mon. This causes a race where mon's threads could access
the now destroyed commandTimeout.

Test Plan:
Run it through the ASAN test with this change
and make sure it passes

Reviewed By: williamspatrick

Differential Revision: D48085535

fbshipit-source-id: 0c1e162a4db5b90f909ce6940bed163f8bccadc9