Set oldest kernel to 5.10
Summary:
https://github.com/yoctoproject/poky/commit/04514378e0136c7d70c20e2a46089353288cedd4
sets up the oldest kernel as 5.15 for all lf-master yocto distros.
As a side-effect glib is built with optimizations which work on those
kernels but might break on older ones.
We know of one change which definitely breaks on 5.0 kernels (glib switches to 64bit time_t which was support added in 5.1)
But 5.6 and 5.10 kernels so far seems to be working fine.

This is more of a safeguard to ensure future optimizations/enhancements do not breaks us. Thus set all kernel 5.10s to
set the OLDEST_KERNEL To 5.10.

Test Plan: TODO

Reviewed By: williamspatrick

Differential Revision: D48199372

fbshipit-source-id: 7329732725aa19bbc579d9028614e11bdcbaacf2