GTI: reverse the logic of the assertion and deassertion for discrete snr
Summary:
# Description

Since the default value of sensor's curr_state is 0. So, when curr_state is 0, indicates the snr is in normal state. We should logging the "deassert" event. On the contrary, when curr_state is 1, indicates the snr is in abnormal state. We should logging the "assert" event,

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4040

Test Plan:
Build and test pass on GTI system

When error occurs:
1    mb       2018-03-09 04:36:06    sensord          ASSERT: MB_RETIMER0_HEALTH - FRU: 1

When error has gone:
1    mb       2018-03-09 05:09:08    sensord          DEASSERT: MB_RETIMER0_HEALTH - FRU: 1

Reviewed By: williamspatrick

Differential Revision: D48422010

fbshipit-source-id: 6a9e36cbb0a87d37d34f220f73fb7fcef62bc134