emmc: switch util-linux RDEPEND to be more explicit
Summary:
The latest refresh of Yocto caused certain FBOSS platforms
to pull in large pieces of util-linux, which in some cases overrides
content from Busybox.  This was due to a general RDEPEND in the
'emmc' MACHINE_FEATURE rather than an explicit RDEPEND on the pieces
of util-linux needed.  Switch to 'util-linux-blkid' since that appears
to be the only part of 'util-linux' which is called from scripts as
part of the eMMC mounting.

Test Plan:
Built fuji and observed much less of util-linux is included
in the image manifest file.  Booted in QEMU and /mnt/data1 was
successfully mounted.

Reviewed By: tao-ren

Differential Revision: D48998231

fbshipit-source-id: fa0b8c89e37425b77cdf56f5782e5b3dd57ef0e7